requirements.txt
flask
openai
python-dotenv
requests
tqdm
markdown
streamlit  
